refactor: jdbc pet repository 예외 사항 추가
findById(), deleteById() 각각 예외 사항 추가
추가된 예외 사항에 대한 테스트 케이스 추가

+++ src/main/java/com/study/db_connection/repository/jdbc/JdbcPetRepository.java
@@ -72,6 +72,9 @@ public class JdbcPetRepository {
     }
 
     public Pet findById(Long id) {
+        if (id == null) {
+            throw new IllegalStateException("id not nullable");
+        }
         String sql = "select * from pet where id = ?";
 
         Connection con = null;
@@ -154,6 +157,9 @@ public class JdbcPetRepository {
     }
 
     public void deleteById(Long id) {
+        if (id == null) {
+            throw new IllegalStateException("id not nullable");
+        }
         String sql = "delete from pet where id = ?";
 
         Connection con = null;
